CARDS-1400: New Questionnaire entry type: Question matrix Moved highlight logic to the Question to avoid code duplication
